+++ simd-minimizers/src/lib.rs
@@ -97,9 +97,20 @@
 //!
 //! let packed_seq = PackedSeqVec::from_ascii(seq);
 //! let mut fwd_pos = Vec::new();
-//! // Unfortunately, `PackedSeqVec` can not `Deref` into a `PackedSeq`.
+//! // Unfortunately, `PackedSeqVec` can not `Deref` into a `PackedSeq`, so `as_slice` is needed.
 //! simd_minimizers::canonical_minimizer_positions(packed_seq.as_slice(), k, w, &mut fwd_pos);
 //! assert_eq!(fwd_pos, vec![3, 5, 12]);
+//!
+//! let mut fwd_vals = Vec::new();
+//! simd_minimizers::extract_canonical_minimizer_values(packed_seq.as_slice(), k, &fwd_pos, &mut fwd_vals);
+//! assert_eq!(fwd_vals, vec![
+//!     // A C G A G, GAGCA is rc of TGCTC at pos 3
+//!     0b0001110011,
+//!     // G A C T C, CTCAG is at pos 5
+//!     0b1100011001,
+//!     // A C T C A, ACTCA is at pos 12
+//!     0b0001100100
+//! ]);
 //!
 //! // Check that reverse complement sequence has minimizers at 'reverse' positions.
 //! let rc_seq = seq.iter().rev().map(|&b| complement_char(b)).collect::<Vec<_>>();
@@ -110,6 +121,10 @@
 //! for (fwd, &rc) in std::iter::zip(fwd_pos, rc_pos.iter().rev()) {
 //!     assert_eq!(fwd as usize, seq.len() - k - rc as usize);
 //! }
+//! let mut rc_vals = Vec::new();
+//! simd_minimizers::extract_canonical_minimizer_values(rc_packed_seq.as_slice(), k, &rc_pos, &mut rc_vals);
+//! rc_vals.reverse();
+//! assert_eq!(rc_vals, fwd_vals);
 //! ```
 #![cfg_attr(
     not(any(
@@ -159,6 +174,8 @@ pub mod private {
     }
     pub use packed_seq::u32x8 as S;
 }
+
+use std::cmp::min;
 
 /// Re-export of the `packed-seq` crate.
 pub use packed_seq;
@@ -261,6 +278,35 @@ pub fn canonical_minimizer_and_superkmer_positions<'s, S: Seq<'s>>(
         let head_tail = canonical_minimizers_seq_simd::<_, MulHasher>(seq, k, w);
         collect_and_dedup_with_index_into(head_tail, min_pos_vec, sk_pos_vec);
     }
+}
+
+/// Given a sequence and a list of positions, extract the k-mer values at those positions.
+pub fn extract_minimizer_values<'s, S: Seq<'s>>(
+    seq: S,
+    k: usize,
+    positions: &[u32],
+    out_vec: &mut Vec<u64>,
+) {
+    out_vec.clear();
+    out_vec.extend(positions.iter().map(|&pos| seq.read_kmer(k, pos as usize)));
+}
+
+/// Given a sequence and a list of positions, extract the *canonical* k-mer values at those positions.
+///
+/// Canonical k-mers are defined as the *minimum* of the k-mer and its reverse complement.
+/// Note that this also works for even `k`, but typically one would want `k` to be odd.
+pub fn extract_canonical_minimizer_values<'s, S: Seq<'s>>(
+    seq: S,
+    k: usize,
+    positions: &[u32],
+    out_vec: &mut Vec<u64>,
+) {
+    out_vec.clear();
+    out_vec.extend(positions.iter().map(|&pos| {
+        let a = seq.read_kmer(k, pos as usize);
+        let b = seq.read_revcomp_kmer(k, pos as usize);
+        min(a, b)
+    }));
 }
 
 /// Variants that always use mulHash, instead of the default ntHash for DNA and mulHash for text.